+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.82
+* Monkey-patch $.fn.dialog only once
+* Leave proper Bootstrap buttons in dialog button sets unchanged
+* Fix "null" fallback text on resultTable buttons
+* Fix initialization integration issues with jqueryui dialog
+* Respect `dblclick` jquerydialogextendjs option if passed in; only default to maximize if option omitted
+
 # 0.2.81
 * Fix copyToClipboard regression
 * Fix checkbox validation state visualization
@@ -8,6 +15,13 @@
 # 0.2.0
 * Drop requirements for assorted IE <= 9 shim scripts only included on demo page
 * Remove digitizing toolset code, this is now moved to the Mapbender Digitizer repository
+
+# 0.1.82
+* Monkey-patch $.fn.dialog only once
+* Leave proper Bootstrap buttons in dialog button sets unchanged
+* Fix "null" fallback text on resultTable buttons
+* Fix initialization integration issues with jqueryui dialog
+* Respect `dblclick` jquerydialogextendjs option if passed in; only default to maximize if option omitted
 
 # 0.1.81
 * Fix copyToClipboard regression